AR-325 (#211)
* AR-325: measure replication delay

* AR-325: measure replication delay

* AR-325: measure replication delay

* AR-325: measure replication delay

* AR-325: measure replication delay

* junits

* AR-325: measure replication delay

* AR-325: measure replication delay

* AR-325: measure replication delay

+++ src/main/java/gov/ca/cwds/neutron/flight/FlightLog.java
@@ -860,30 +860,30 @@ public class FlightLog implements ApiMarker, AtomRocketControl {
       attribs.putIfAbsent("run_start_time", Instant.ofEpochMilli(startTime).getEpochSecond());
       attribs.putIfAbsent("run_end_time", Instant.ofEpochMilli(endTime).getEpochSecond());
 
-      final long runSeconds = (endTime - startTime) / 1000;
+      final float runSeconds = (endTime - startTime) / 1000;
       LOGGER.debug("Neutron: this run seconds: {}", runSeconds);
       attribs.putIfAbsent("run_seconds", runSeconds);
 
-      final long runMillis = endTime - startTime;
+      final float runMillis = endTime - startTime;
       attribs.putIfAbsent("run_millis", runMillis);
 
       // AR-325: replication metrics.
       // blue line: DB2 replication.
-      // green line: Neutron processing time + ES refresh interval.
+      // green line: job processing time + ES refresh interval + delay between job runs.
       if (lastEndTime != 0) {
         attribs.putIfAbsent("last_run_end_time",
             Instant.ofEpochMilli(lastEndTime).getEpochSecond());
-        final long runTotalMillis = lastEndTime - startTime;
-        final long runTotalSeconds = runTotalMillis / 1000;
+        final float runTotalMillis = lastEndTime - startTime;
+        final float runTotalSeconds = runTotalMillis / 1000;
 
         LOGGER.debug("since last run: millis: {}, seconds: {}", runTotalMillis, runTotalSeconds);
         attribs.putIfAbsent("run_since_last_run_secs", runTotalSeconds);
         attribs.putIfAbsent("run_since_last_run_millis", runTotalMillis);
 
-        final long totalGreenLineSecs = runTotalSeconds + 3; // NEXT: ES refresh interval
+        final float totalGreenLineSecs = runTotalSeconds + 3; // NEXT: ES refresh interval
         attribs.putIfAbsent("green_line_secs", totalGreenLineSecs);
 
-        final long totalGreenLineMillis = runTotalMillis + 3000; // NEXT: ES refresh interval
+        final float totalGreenLineMillis = runTotalMillis + 3000; // NEXT: ES refresh interval
         attribs.putIfAbsent("green_line_millis", totalGreenLineMillis);
       }
